requirements.txt
# Core Dependencies
openai>=1.40.0
tiktoken>=0.5.0
tenacity>=8.2.0
plotly>=5.20.0
kaleido>=0.2.1  # For static chart export in PDF
pandas>=2.2.0
numpy>=1.26.0
pyyaml>=6.0.1
jinja2>=3.1.3
python-dateutil>=2.9.0
openpyxl>=3.1.2
aiohttp>=3.9.0
python-dotenv>=1.0.0
markdown>=3.5.0
beautifulsoup4>=4.12.0
nest-asyncio>=1.5.0  # For handling async event loops in Jupyter/async contexts

# PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.9.0

# Vector Database & Embeddings
faiss-cpu>=1.7.4

# Web Search (Optional - choose one)
tavily-python>=0.3.0  # Alternative to OpenAI browsing

# PDF Export
weasyprint>=60.0
playwright>=1.40.0  # For chart screenshots in PDF

# LangChain for Enhanced Features
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
langchain-community>=0.2.0

# Additional utilities
chardet>=5.2.0  # For encoding detection
python-docx>=1.1.0  # For DOCX export
pypandoc>=1.11  # For document conversion
markdown-it-py>=3.0.0  # Enhanced markdown processing
streamlit>=1.36.0  # Web UI

# Logging and Monitoring
colorlog>=6.7.0  # Colored logging output

# Development & Testing (Optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0  # Code formatting
flake8>=6.0.0  # Linting